+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 ### Added
 - `CellFilter::NonEmpty`: selects cells that contain a non-space symbol.
+- `Interpolation::SmoothStep` and `Interpolation::Spring`.
 
 ## tachyonfx-0.24.0 - 2026-02-14
 ### Added